fix scale bug
fix bug with scale

+++ io_scene_4a_engine/match.py
@@ -127,6 +127,7 @@ class UVec4S16:
 
         return UVec4S16(x, y, z, w)
 
+
 class Vec4S16:
     X: int = 0
     Y: int = 0
@@ -140,10 +141,10 @@ class Vec4S16:
         self.W = w
 
     def division(self, div_kef):
-        self.X = int(self.X / div_kef)
-        self.Y = int(self.Y / div_kef)
-        self.Z = int(self.Z / div_kef)
-        self.W = int(self.W / div_kef)
+        self.X = self.X / div_kef
+        self.Y = self.Y / div_kef
+        self.Z = self.Z / div_kef
+        self.W = self.W / div_kef
 
     @staticmethod
     def read(rd: Reader):
